[feat] 장바구니에 담긴 물품 조회

+++ src/main/java/com/example/shop/domain/product/Product.java
@@ -6,6 +6,8 @@ import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Entity
 @Getter
 @NoArgsConstructor
@@ -20,9 +22,9 @@ public class Product  extends BaseEntity {
     @Column(name = "product_name")
     private String productName;
     @NotNull
-    private Long quantity;
+    private int quantity;
     @NotNull
-    private Long price;
+    private BigDecimal price;
     private String image;
     private String description;
 }
